backend/worker/requirements.txt
flask
gunicorn
supabase
google-generativeai
langchain-core
langchain-google-genai
pypdf